+++ resources/templates/ts/ui.dsl.layout.ts
@@ -15,13 +15,15 @@ function createHtml(): string {
         ui.input(
           { id: 'search', placeholder: 'Search' },
           ui.onHandle('input', ({ value }) => {
+            // This logs to the Developer Tools console
             console.log('Input changed:', value);
           }),
+          ui.button({ id: 'caseButton', label: 'Aa', toggle: true }),
         ),
         ui.button(
           { id: 'searchButton', label: 'Search' },
-          ui.on('click', 'onSearch'),
-          ui.handler('onSearch', () => {
+          ui.onHandle('click', () => {
+            // This logs to the Developer Tools console
             console.log('Search clicked');
           }),
         ),
@@ -29,7 +31,7 @@ function createHtml(): string {
       ui.tree(
         {
           id: 'exampleTree',
-          enableRemove: true,
+          remove: true,
           initialItems: [
             {
               id: 'root',
@@ -39,6 +41,7 @@ function createHtml(): string {
           ],
         },
         ui.onHandle('activate', ({ item }) => {
+          // This logs to the Developer Tools console
           console.log('Node activated:', item);
         }),
       ),